+++ tanjun/annotations.py
@@ -193,25 +193,54 @@ User = typing.Annotated[hikari.User, _OptionMarker(hikari.User)]
 """An argument which takes a user."""
 
 
+class _FloatEnumConverter(_ConfigIdentifier):
+    """Specialised converters for float enum choices."""
+
+    __slots__ = ("_enum",)
+
+    def __init__(self, enum: collections.Callable[[float], typing.Any]) -> None:
+        self._enum = enum
+
+    def set_config(self, config: _ArgConfig, /) -> None:
+        config.float_converter = self._enum
+
+
+class _IntEnumConverter(_ConfigIdentifier):
+    """Specialised converters for int enum choices."""
+
+    __slots__ = ("_enum",)
+
+    def __init__(self, enum: collections.Callable[[int], typing.Any]) -> None:
+        self._enum = enum
+
+    def set_config(self, config: _ArgConfig, /) -> None:
+        config.int_converter = self._enum
+
+
 class _ChoicesMeta(abc.ABCMeta):
     def __getitem__(cls, enum_: type[_EnumT], /) -> type[_EnumT]:
-        if issubclass(enum_, int):
+        if issubclass(enum_, float):
+            type_ = float
+            choices = Choices(enum_.__members__)
+            converter = _FloatEnumConverter(enum_)
+
+        elif issubclass(enum_, int):
             type_ = int
             choices = Choices(enum_.__members__)
+            converter = _IntEnumConverter(enum_)
 
         elif issubclass(enum_, str):
             type_ = str
             choices = Choices(enum_.__members__)
-
-        elif issubclass(enum_, float):
-            type_ = float
-            choices = Choices(enum_.__members__)
+            converter = None
 
         else:
             raise TypeError("Enum must be a subclass of str, float or int")
 
         # TODO: do we want to wrap the convert callback to give better failed parse messages?
-        return typing.cast(type[_EnumT], typing.Annotated[enum_, choices, Converted(enum_), _TypeOverride(type_)])
+        return typing.cast(
+            type[_EnumT], typing.Annotated[enum_, choices, converter, Converted(enum_), _TypeOverride(type_)]
+        )
 
 
 class Choices(_ConfigIdentifier, metaclass=_ChoicesMeta):
@@ -326,7 +355,8 @@ class Converted(_ConfigIdentifier, metaclass=_ConvertedMeta):
         return self._converters
 
     def set_config(self, config: _ArgConfig, /) -> None:
-        config.converters = self._converters
+        config.str_converters = self._converters
+        config.option_type = str
 
 
 Color = typing.Annotated[hikari.Color, Converted(conversion.to_color)]
@@ -528,9 +558,6 @@ class Positional(_ConfigIdentifier, metaclass=_PositionalMeta):
     """
 
     __slots__ = ()
-
-    def __init__(self) -> None:
-        """Create a positional instance."""
 
     def set_config(self, config: _ArgConfig, /) -> None:
         config.is_positional = True
@@ -1097,10 +1124,11 @@ class _ArgConfig:
         "aliases",
         "channel_types",
         "choices",
-        "converters",
         "default",
         "description",
         "empty_value",
+        "float_converter",
+        "int_converter",
         "is_greedy",
         "is_positional",
         "min_length",
@@ -1113,16 +1141,19 @@ class _ArgConfig:
         "range_or_slice",
         "slash_name",
         "snowflake_converter",
+        "str_converters",
     )
 
     def __init__(self, parameter: inspect.Parameter, /, *, description: typing.Optional[str]) -> None:
         self.aliases: typing.Optional[collections.Sequence[str]] = None
         self.channel_types: typing.Optional[collections.Sequence[_ChannelTypeIsh]] = None
         self.choices: typing.Optional[collections.Mapping[str, _ChoiceUnion]] = None
-        self.converters: collections.Sequence[_ConverterSig[typing.Any]] | None = None
         self.default: typing.Any = parsing.UNDEFINED if parameter.default is parameter.empty else parameter.default
         self.description: typing.Optional[str] = description
         self.empty_value: typing.Union[parsing.UndefinedT, typing.Any] = parsing.UNDEFINED
+        self.float_converter: collections.Callable[[float], typing.Any] | None = None
+        self.int_converter: collections.Callable[[int], typing.Any] | None = None
+        # The float and int converters are just for Choices[Enum].
         self.is_greedy: bool = False
         self.is_positional: typing.Optional[bool] = None
         self.min_length: typing.Optional[int] = None
@@ -1135,6 +1166,7 @@ class _ArgConfig:
         self.range_or_slice: typing.Union[range, slice, None] = None
         self.slash_name: str = parameter.name
         self.snowflake_converter: typing.Optional[collections.Callable[[str], hikari.Snowflake]] = None
+        self.str_converters: collections.Sequence[_ConverterSig[typing.Any]] | None = None
 
     def finalise_slice(self) -> None:
         if not self.range_or_slice:
@@ -1158,8 +1190,8 @@ class _ArgConfig:
             self.max_value = max_value
 
     def to_message_option(self, command: message.MessageCommand[typing.Any], /) -> None:
-        if self.converters:
-            converters = self.converters
+        if self.str_converters:
+            converters = self.str_converters
 
         elif self.option_type:
             if self.snowflake_converter and self.option_type in _MESSAGE_ID_ONLY:
@@ -1225,15 +1257,14 @@ class _ArgConfig:
         return slash.UNDEFINED_DEFAULT if self.default is parsing.UNDEFINED else self.default
 
     def to_slash_option(self, command: slash.SlashCommand[typing.Any], /) -> None:
-        option_type = self.option_type
-        if not option_type and self.converters:
-            option_type = str
+        if self.str_converters and not self.option_type:
+            self.option_type = str
 
-        if option_type:
+        if self.option_type:
             if not self.description:
                 raise ValueError(f"Missing description for argument {self.parameter.name!r}")
 
-            self.SLASH_OPTION_ADDER[option_type](self, command, self.description)
+            self.SLASH_OPTION_ADDER[self.option_type](self, command, self.description)
 
     SLASH_OPTION_ADDER: dict[
         typing.Any, collections.Callable[[Self, slash.SlashCommand[typing.Any], str], slash.SlashCommand[typing.Any]]
@@ -1251,6 +1282,7 @@ class _ArgConfig:
             self.slash_name,
             d,
             choices=_ensure_values("choice", float, self.choices),  # TODO: can we pass ints here as well?
+            converters=self.float_converter or (),
             default=self._slash_default(),
             key=self.parameter.name,
             min_value=self.min_value,  # TODO: explicitly cast to float?
@@ -1260,6 +1292,7 @@ class _ArgConfig:
             self.slash_name,
             d,
             choices=_ensure_values("choice", int, self.choices),
+            converters=self.int_converter or (),
             default=self._slash_default(),
             key=self.parameter.name,
             min_value=_ensure_value("min", int, self.min_value),
@@ -1278,7 +1311,7 @@ class _ArgConfig:
             self.slash_name,
             d,
             choices=_ensure_values("choice", str, self.choices),
-            converters=self.converters or (),
+            converters=self.str_converters or (),
             default=self._slash_default(),
             key=self.parameter.name,
             min_length=self.min_length,
@@ -1369,7 +1402,7 @@ def parse_annotated_args(
                 arg_config.range_or_slice = arg
 
         arg_config.finalise_slice()
-        if arg_config.option_type or arg_config.converters:
+        if arg_config.option_type or arg_config.str_converters:
             for slash_command in slash_commands:
                 arg_config.to_slash_option(slash_command)